requirements.txt
pandas==1.5.2
pyspark==3.4.0
python-dotenv==1.0.0
Requests==2.29.0
tqdm==4.64.1
